finalChanges2

+++ Binomial.java
@@ -23,7 +23,7 @@ public class Binomial {
 	
 	// Computes the Binomial function, efficiently
 	public static int binomial(int n, int k) {
-		int[][] memo = new int[n + 1][k + 1];
+		long[][] memo = new long[n + 1][k + 1];
 
     	for (int i = 0; i <= n; i++) {
         	for (int j = 0; j <= k; j++) {
@@ -34,7 +34,7 @@ public class Binomial {
     return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (k > n) {
         	return 0;
     }
